Changed contract and tests to mapping
