+++ test/helpers/transposeArray.js
@@ -1,0 +1,10 @@
+const tranposeArray = array => {
+  return array[0].map((col, i) => {
+    return array.map(row => row[i])
+  })
+}
+
+module.exports = {
+  tranposeArray
+}
+
